Token verification

+++ routes/transactionRoutes.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import { POSSIBLE_CATEGORIES, POSSIBLE_SECTORS, createTransaction, deleteTransaction, listTransactions, validateCategory, validateSector } from "../controllers/transactionController";
-import jwt from "jsonwebtoken"
+import jwt from 'jsonwebtoken';
 
 export const transactionRouter = express.Router()
 
@@ -22,39 +22,48 @@ function parseCookies (request) {
   return list;
 }
 
-transactionRouter.use((req, res, next) => {
-  const cookiesList = parseCookies(req)
-  const decoded = jwt.decode(cookiesList["session"], {complete: true});
-  if (decoded === null){
-    return res.status(401).send({detail: "Invalid token"});
-  }
-  next()
-})
+
 
 
 transactionRouter.post("/", async (req, res) => {
-  if (!validateSector(req.body.sector))
-    res.status(400).send({detail: `Invalid sector, please insert a sector that is part of ${POSSIBLE_SECTORS}`})  
-  
-  if(!validateCategory(req.body.category))
-    res.status(400).send({detail: `Invalid category, please insert a category that is part of ${POSSIBLE_CATEGORIES}`})  
-
-  const cookieList = parseCookies(req)
-  const username = cookieList["username"]
-  
-  res.send(await createTransaction(username, req.body.value, req.body.category, req.body.sector))
+  if (!req.headers.authorization) {
+    return res.status(403).json({ error: 'No credentials sent!' });
+  }
+  jwt.verify(req.headers.authorization.replace('Bearer ',''), process.env.TOKEN_SECRET, async function(error, decoded){
+    if(error)
+      res.status(403).send({detail: "Invalid or expired token"});
+    
+    if (!validateSector(req.body.sector))
+      res.status(400).send({detail: `Invalid sector, please insert a sector that is part of ${POSSIBLE_SECTORS}`})  
+    
+    if(!validateCategory(req.body.category))
+      res.status(400).send({detail: `Invalid category, please insert a category that is part of ${POSSIBLE_CATEGORIES}`})  
+    
+    res.send(await createTransaction(decoded.username, req.body.value, req.body.category, req.body.sector))
+  });
 });
 
 transactionRouter.get("/", async (req, res) => {
-  const cookieList = parseCookies(req);
-  const username = cookieList["username"];
-
-  res.send(await listTransactions(username));
+  if (!req.headers.authorization) {
+    return res.status(403).json({ error: 'No credentials sent!' });
+  }
+  jwt.verify(req.headers.authorization.replace('Bearer ',''), process.env.TOKEN_SECRET, async function(error, decoded){
+    if(error)
+      res.status(403).send({detail: "Invalid or expired token"});
+    res.status(200).send(await listTransactions(decoded.username));
+  });
 })
 
 transactionRouter.delete("/:transactionId", async (req, res) => {
-  const deletedTransaction = await deleteTransaction(req.params.transactionId);
-  if(deleteTransaction)
-    res.status(204).send();
-  res.status(404);
-})
+  if (!req.headers.authorization) {
+    return res.status(403).json({ error: 'No credentials sent!' });
+  }
+  jwt.verify(req.headers.authorization.replace('Bearer ',''), process.env.TOKEN_SECRET, async function(error, decoded){
+    if(error)
+      res.status(403).send({detail: "Invalid or expired token"});
+      const deletedTransaction = await deleteTransaction(req.params.transactionId);
+      if(deleteTransaction)
+        res.status(204).send();
+      res.status(404);
+  });
+});
